starting to restore view funcitonality

+++ lib/io/basic_buffer.cpp
@@ -4,13 +4,13 @@ namespace tachyon{
 namespace io{
 
 void SerializeString(const std::string& string, io::BasicBuffer& buffer){
-	size_t size_helper = string.size();
+	uint32_t size_helper = string.size();
 	buffer += size_helper;
 	buffer += string;
 }
 
 void DeserializeString(std::string& string, io::BasicBuffer& buffer){
-	size_t size_helper;
+	uint32_t size_helper;
 	buffer >> size_helper;
 	string.resize(size_helper);
 	buffer.read(&string[0], size_helper);
